Initial commit
This is the initial commit for the 'merge' of the backend and the front end of the soil correction application, the pom.xml was changed to have the dependencies for both projects. the files had their current package changed and the only file that was changed beyond that was NutrientesCTC because the GUI project was made initially in java 14 which does not have the record keyword available, the changed made it so the use of the class is still the same. the backend will still be organized in a different file structure